Fin du projet

+++ controller/ControllerCours.php
@@ -71,12 +71,7 @@ class ControllerCours extends Controller{
 		$cours=new ModelCours($data);
 		$cours->save($data);
 
-		$tab = ModelCours::selectAll();
-		$path=array('model','ModelMatieres.php');
-		require_once File::build_path($path);
-		$view="list";
-		$pagetitle="Tous mes cours - Agora";
-		require(File::build_path(array('view','view.php')));
+		header('Location: ./index.php?controller=cours&action=list');
 
 		
 	}
@@ -84,9 +79,26 @@ class ControllerCours extends Controller{
 
 	public static function suppr(){
 
-		$codeCours = $_GET["code"];
-		ModelCours::delete($codeCours);
-		self::list();
+		if(Session::is_teacher()){
+
+			$codeCours = $_GET["code"];
+			ModelCours::delete($codeCours);
+
+			if(isset($_GET['redirection'])){
+
+				header('Location: ./index.php?controller=enseignants&action=show_perso_page');
+
+			}else{
+
+				self::list();
+
+			}
+
+		}else{
+
+			header('Location: ./index.php?controller=Utilisateurs&action=show_login_page');
+
+		}
 
 
 	}
